Use xvfb-run on Travis and fix datetime issues
xvfb allows capybara-webkit to work on Travis.

Rails is set to use the Australia/Melbourne timezone and Travis was not,
this caused failing tests on Travis.
The solution is to set Chronic to use the ActiveSupport's timezone
instead of the defaulting to the system's timezone.

For more info: https://github.com/mojombo/chronic#time-zones

Just seeing if the in_time_zone actually helped