Use short URLs for users and long URLs for admins
This means you should be able to safely give people the standard URL to
an event and they won't be able to modify anything, unless they know the
other, longer URL.

These URLs are not stored in the database, they are based on a
hash of the id (Hashids) with a given salt.

Unfortunately, this also means that the salt can't be changed easily
while people are using the site, as nothing is stored.
A possible way to mitigate this would be to use the same code to
generate the two URLs, but store them in the database.
This means that all previous URLs can be preserved while any new ones
use a new (possibly random) salt.
It could also allow an event admin to revoke and regenerate new URLs.